sockpuppet-audit/requirements.txt
playwright==1.47.0
pandas==2.2.2
pyyaml==6.0.2
